Move existing documentation to org site
This is in preparation for adding documentation for multiple Millennium
Machines projects, which would not make sense being in the Milo
repository.

This commit splits out the existing documentation setup from the Milo
repo, and reconfigures it to work correctly in the organisation
repository.

It also implements deployment of a dev site, so changes can be tested
against a `dev` branch before merging back into the `main` branch.

Signed-off-by: Ben Agricola <ben+git@agrico.la>