Add Toolsetter guide and restructure for multi doc
This PR splits the documentation out further into project-related
sections - one for Milo and one for the Toolsetter (Long John), allowing
us further scope to expand it to other projects like Casa and Fixated in
future.

It also fixes various spelling and grammar issues, and consistently refers
to threadlock rather than loctite or thread-lock or similar.

Signed-off-by: Ben Agricola <ben+git@agrico.la>